aquaregia: Expect error count via len(name slice)
Instead of calculating the expected number of errors per file based on the
repeated counts of the same error case struct, I instead read the length of a
TestFuncs slice to determine how many times a match error is expected to
occur.

This de-duplicates test structs while also programatically enforcing that
the test code identifies where the expected failures are supposed to happen.

I've also added a check in the test to ensure that failure struct cases
are unqiue.

Initial idea for this in feedback for D6722147.

Refs GO-950
Reviewed-on: D6765303