parallel cleanup: Remove job field
The task struct for cff.Parallel does not require a job field because
cff.Parallel tasks are independent. Under parallel, enqueues to the CFF
scheduler do not require dependency jobs to be set nor does the job returned
by the scheduler enqueue to need to be stored in any field.

The task struct is still useful to keep around for the purposes of
tracking state to determine whether or not a parallel task was run or
skipped. This change will be added in a subsequent diff.

Refs GO-84
Reviewed-on: D6778415