cleanup: Test InstrumentParallel with Tasks
Fixes a bug where a test Parallel contains a cff.Task option. At the point
in the stack cff.Task for Parallel has not yet been introduced. However this
does not error as cff.Task is a real option but is ignored by the code that
compiles cff.Parallel.

The change replaces cff.Task with cff.Tasks (which does exist for
cff.Parallel at this point in the stack). The results of the test are
unchanged as the error case the test is expecting happens before cff.Tasks
or cff.Task(s) would be processed.

Refs GO-84
Reviewed-on: D6795257