message: Reference gazelle and bazel
Now that the CFF gazelle extension is online, we can update the
non-generated code panic messaging with more helpful instructions.

Users will likely not know what it means to 'run cff' since the CFF code
generation CLI is not something we expect users to know or care about.
In this light, replace that with a message to direct users to run
'gazelle' and 'bazel', which are commands all go developers will know about.

Test Plan: {F571093197}

Refs GO-1003
Reviewed-on: D6899469