Fix variable shadowing collisions in cff.Flow and cff.Predicate
If a user provides a function like

```
func Something(ctx thrift.Context) {
	// ...

	cff.Flow(
		...,
		cff2.Task(func() {
			doSomething(ctx)
		}
	)
}
```

The `ctx` in the function frame should point to the `ctx thrift.Context`
in the arguments above. However, the CFF generated code has a
`ctx context.Context` argument that shadows this and causes a
compilation error from type-mismatch. While the current shadowing
happens because of the `ctx` variable, this can happen for any variable.

This commit fixes these issues by evaluating the functions first, then
generating the functions necessary for CFF. By doing so, the functions
evaluated point to the correct variables and removes variable shadowing.

If a user inspects the generated code, they'll notice that the variables
are weirdly shaped: something like `_15_12`. This denotes the line and
column numbers of where the function is written in the original code.
This is done because:

- We need a unique ID for each cff.Task and Predicate function, but we
cannot carry over the ID values between templates, so we needed to
create some unique ID using the AST node of the function itself.
- This ensures that there are no variable conflicts: The line and column
number is unique to each Task or Predicate function, and each CFF
generated code for a single cff.Flow is isolated in a function.

Note:
- This diff introduces an issues with parameter ordering, the fix for
  this issue is stacked above (D7090423).
- Since @minho.park has left Uber, i've (rhang@) taken over the diff
  to address the code review comments.

Resolves GO-1098

Reviewed-on: D6954889