[go/importer] Remove unnecessary parameters from LoadPackages
LoadPackages currently takes in a LoadParams struct, even though
only the ImportPath is used from it. This was a remnant from the
original cff importer that this library is based off, and it doesn't
really make sense to do this.

This changes LoadPackages to directly take in the import path as a
string type.

Refs GO-935

Reviewed-on: D7268461