Fix cff shadowed variables in cff.Parallel,cff.Map, and cff.Slice
This addresses issues with variable shadowing in cff.Parallel, cff.Map,
and cff.Slice. For example, when cff.Parallel (or cff.Map or
cff.Slice) is used like this:

```
func SomeFunc(ctx SomeContextType) {
	cff.Parallel(
		// ...,
		cff.Task(func() {
			doSomething(ctx)
		},
		// ...
	)
}
```

The `ctx` variable gets overwritten with `context.Context`, which
results in unintended behavior.

This fix uses the same approach as the one used in D7090423.

Refs GO-1099

Reviewed-on: D7374321