scheduler: Remove waiting queue
In CFF scheduler, the `waiting` queue is actually not really doing much
other than bookkeeping how many jobs are waiting to be executed because
they depend on jobs that are still running or haven't executed yet.

Currently, the jobs are either placed in the `waiting` queue or the `ready`
queue, and have to maintain an index in the waiting queue when they are
in the waiting queue because removing from an arbitrary index is cheaper
that way.

All these can be eliminated by removing the `waiting` queue and simply
keeping track of the number of items waiting. Jobs that need to be
changed to `ready` state is done so when jobs that they depend on has
finished executing, so no extra bookkeeping of all jobs in `waiting`
state is necessary.

Refs GO-1323

Reviewed-on: D7376111