Detect unexpected worker termination in CFF scheduler
CFF worker goroutines run arbitrary user code, which can call code that
unexpectedly kills the goroutine. This may lead to problem as the
scheduler may run out of worker goroutines if sufficient number of such
calls are made.

This changes the CFF scheduler to detect any unexpected termination
in the worker goroutines. If the worker thread was terminated
unexpectedly, that is reported as part of the Job failure error, and
the scheduler respawns the worker goroutine that was killed.

Refs GO-1100

Reviewed-on: D7375275