Support for wrapper types in cff.Slice
Currently, cff.Slice does not allow wrapper types for slices to be
passed as an argument to cff.Slice. For example:

type stringList []string

is not a valid argument to cff.Slice.

This extends the type checking done in compileParallel to allow for such
types by inspecting the underlying type for the argument type, and only
erroring out if both the given type and the underlying type are not
slice types.

Refs GO-1005

Reviewed-on: D7441237