Fix a flaky CFF2 parallel test
This fixes a flaky test in CFF2. Specifically, the tests that make
use of AssignSliceItems to error out or panic during the middle of
execution and asserting on the content of the slice can run into
a race.

This occurs because the goroutine that was executing on an element
that makes it panic or exit with an error makes the function to exit
and the main goroutine tries to compare the target slice in the test
whereas the rest of the goroutines working on other elements of the
slice are not yet done and are still writing to it. This causes the
race detector to sometimes go off.

The fix is changing AssignSliceItems to always assign items to the
corresponding target slice before panicking or erroring out, and
making the assertion directly on the element which is guaranteed
to have error/panic. We shouldn't be asserting on elements that are
not guaranteed to have been finished.

Refs GO-1345.

Test Plan:
Tested with 1000 iterations of the test and verified no tests fail:
```
[sungyoon@sungyoon]..ber.internal/marketplace/cff❯ bazel test --runs_per_test=1000 --cache_test_results=no //src/go.uber.org/cff/internal/tests/parallel/...
INFO: Invocation ID: 35903599-82dc-4958-a92d-12123a7041d2
INFO: Build option --test_env has changed, discarding analysis cache.
INFO: Analyzed 3 targets (218 packages loaded, 11784 targets configured).
INFO: Found 2 targets and 1 test target...
INFO: Elapsed time: 16.885s, Critical Path: 1.33s
INFO: 1001 processes: 1 internal, 1000 processwrapper-sandbox.
INFO: Build completed successfully, 1001 total actions
//src/go.uber.org/cff/internal/tests/parallel:go_default_test PASSED in 1.1s
  Stats over 1000 runs: max = 1.1s, min = 0.2s, avg = 0.6s, dev = 0.2s
```

Reviewed-on: D7455937