Fix reset line directive
The line directives generated in CFF2 maps the lines to either:
1) The original source
2) The generated source

Lines that are user-written are mapped to the corresponding line in
the original source.

Lines that are part of the CFF-generated code do not have corresponding
lines in the original source, so they are mapped to their own lines.

This fixes the second type of the line directives: the ones that map the
lines to the file itself - the "reset" directives. Specifically, these
were of the form ":300" which drops the filename.

But the Go compiler parses that as-is, and reports errors like this upon
compilation failure:

.go:276: too many arguments in call to _42_4
        have (context.Context, int, string)
        want (context.Context, string)

This fixes the generated line directive to be of the format:
"file_gen.go:300" instead, which fixes the compilation error messages to
be like the following:

parallel_gen.go:276: too many arguments in call to _42_4
        have (context.Context, int, string)
        want (context.Context, string)

Refs GO-1484.

Reviewed-on: D7772515