test race: Replace map with slice
A test case concurrently modifies a map within a call to cff.Map. The
standard map type in Go is not concurrent safe. Setting the values of
the map are not critical to the intent of the test, which is to prove
that variables are not shadowed.

I've replaced the map logic with a slice for which indicies of the slice are
not concurrently modified.

I've verified the fix with
`bazel test --runs_per_test=1000 --cache_test_results=no :go_default_test`
on the affected package.

Refs GO-618
Reviewed-on: D7814253