cff.Slice: make index parameter optional
This adds the possibility for `cff.Slice` to receive a function with one non-context
parameter (the element of the slice) making the index parameter optional. This helps
the user to write less code and also avoid confusion in cases like when `cff.Slice`
receives an slice of integers, making clear that the only non-context value is the
element of the slice.

Refs GO-1007
Reviewed-on: D7813111