Skeleton code for codegen V2
This is the initial diff for "codegen v2" for CFF2. This is a
currently WIP code generator that we are hoping to implement as
part of the code coverage effort for CFF where we generate code
that do not hinder the line numbers of the existing code.

This generator is not to be used in production until it is deemed
feature parity with the existing code generator, and we will "flip"
the switch once that happens, and remove the existing code generator
after sufficient verification from production uses of the new code
generator. It is currently guaraded under a newly added flag, which
is not added to the Bazel rule (because we don't want anybody using
it yet).

In this diff, we introduce the concept of a modifier -- conceptually,
a modifier is a source modifier that maps 1:1 to a cff "public API"
surface that users can call. For instance, cff.Flow translates to a
flow modifier.

A modifier does two things:
1. It generates a unique per-call function signature that matches the
parameters provided by the user (and its implementation).
2. It inline-replaces the API call with a call to the generated function.

Since each modifier is unique to each call, we're currently using the
line/col number to distinguish them, but that may change in the future
using randomized tokens.

In addition, the current code generated by v2 codegen does not compile
because of type mismatches in parameters. This will be addressed in
follow-up diffs that add corresponding modifiers to each cff primities
or directives. Currently in this diff there are only concurrencyModifier
and a partial implementation of flowModifier.

In addition, I will also follow up with a unit test for modifiers as well
as end-to-end testing for v2 codegen.

Refs GO-1265.

Reviewed-on: D8044671