modifiers: Fix too many cff.Results modifiers bug
A cff.Results modifier should only be produced once per cff.Results
directive in the CFF source code.

Move cff.Results modifier production outside of the "Results" processing
for loop in the CFF compile step so that this is the case.

A test condition is is added to verify that only one cff.Results is
produced despite multiple cff.Results arguments.

Refs GO-1585
Reviewed-on: D8267203