Fix collision issue in modifier generated func names
When modifiers generate functions, it uses the corresponding
cff primitives' positions in their respective files to generate
the names. However, this alone cannot guarantee uniqueness in the
package because it's possible that there may be identical primitives
located in the same line/col pairs.

This fixes that issue by adding the file name to the generated
functions, and verifies that with a test.

Refs GO-1675
Reviewed-on: D8274073