examples: Enable v2 test, better instructions
Enable the test that verifies that magic_v2_gen is also up-to-date.
To do this, add a new cff_v2 target that builds magic_v2
with the modifiers genmode.

Further, the instructions to run the cff CLI have been inaccurate
for a while.
Replace those instructions with a suggestion to run bazel build
and copy the generated file like that.

Refs GO-1684

Reviewed-by: Sung Yoon Whang <sungyoon@uber.com>
Reviewed-by: moises <moises@uber.com>
Reviewed-on: D8457085

Reviewed-on: D8457085