Add back support for "cff" build tag
This adds back support for the `+build cff` tag,
with updates for the new `//go:build` directive.
With this change, if either tag is present,
the generated file will contain an inversion of that.
If the tag is absent, we won't do anything differently.

Threfore, this covers both cases:
Bazel usage, where the tag should be absent,
and non-Bazel usage where the tag should be present.

Updates to magic_gen and magic_v2_gen:
They get the inversions of the tags and all lines get offset by three
because of the two lines for the tags and the newline afterwards.

Resolves GO-1684

Reviewed-on: D8458513