Initialize Go module
Initialize a new module with the import path go.uber.org/cff,
and run go mod tidy.

Note that this introduces an unnecessary dependency on gofrs/uuid
which we should aim to drop in a future change.
We'll likely want to create unpublished submodules for some of the test
cases to reduce dependency bloat.

Refs GO-269

Reviewed-on: D8487635