tests: Fix aquaregia_test failure tests
aquaregia_test takes several example files that fail to compile,
and verifies that the error messages we see when we compile them are
what we expect.

To prevent these files from being touched by our regular toolchain,
we tag them with "failing",
and aquaregia_test opts into this build constraint.

Note that this also deletes the redundant examples/cycle.go file which
we did not use for any other test as aquaregia_test already covers
cycles.

Refs GO-269

Reviewed-on: D8487651