pkg/goPackages: Support build tags
This adds support in the go/packages loader to accept CLI flags
which specify other build tags that must be enabled in addition to 'cff'
for the files it's looking at.

For example,

    cff -tag foo ./...

This will process only files tagged with "cff" and "foo".

Besides providing flexibility to users,
with this we can generate code for different parts of the cff examples
with different cff build flags.

Refs GO-269

Reviewed-on: D8487631