test/examples: Fix clean test
magic_clean_test fails because of two reasons:

- code generated by gen v2 does not fully compile yet
- it relies on Bazel to provide the generated file to compare

To address the first issue, tag the v2 files with the 'v2' constraint.
We opt into that constraint while generating the file.

To address the second issue, we'll shell out to "cff" to generate the
file, and then compare the output.
Note that this means we cannot run `go test` inside this directory
without installing cff globally or without,

    PATH=$(pwd)/../bin:$PATH go test

We can address this in a future change by making examples call the
compiler directly like aquaregia_test.

Refs GO-269

Reviewed-on: D8487671