docs: Minimal docs website
This adds a basic docs website for cff similar to Fx
with the following high-level information:

- A short tutorial that demonstrates use of cff.Flow.
  For lack of enough readily available but sufficiently complex
  third-party APIs,
  I elected to come up with a fake "uber service"
  and a fake client implementation for it.
- An Introduction that goes into use-cases and non-use-cases.
- Installation and setup instructions for projects.
- Concepts: the high level concepts defined by cff
    - Architecture: overview of how it all works
- FAQ: I expect this to grow in the future.
- Contributing and changelog

As with Fx, this uses mdox to pull code snippets for examples,
although not all code samples are currently backed by real snippets --
only the tutorial.

Finally, this includes a GitHub workflow to publish these docs to
uber-go.github.io/cff when we open the repository.