compile: Make 'cff' constraint required
This adds a new option to the compiler which, when enabled,
checks that the 'cff' build constraint is present in a file.

The option is always enabled for the 'cff' binary,
but we can disable it for Bazel the same way we install
a different package loader.

This will ensure that someone does not attempt to use cff code
generation directives outside a file that isn't tagged 'cff',
because if they do that they'll end up with a panicking binary
or code that fails to compile (because of duplicate declarations).

Note that we only check for the *presence* of the tag,
not whether it evaluates to true.
We can assume that it evaluates to true because the package loader
always passes that in with `-tags cff`.

Resolves #4