docs/mdox: Fix GitHub regex
The githubPullsIssues validator (which uses GitHub API to validate links)
does not yet support discussions, so it refuses that regular expression.

Pull out the discussions part of the regex into a separate validator
entry that we mark as ignored.

Because otherwise, it would hit GitHub for each discussion we ever link
to separately and run into rate limiting issues.