docs(get-started): Slightly more demonstrative example
This changes the get-started tutorial to be slightly more demonstrative
of the value of cff.

It does so by introducing another resource: location,
and adding a call to it when one of the types are available.

So instead of a simple fan out and fan back in,
you get a graph where one of the nodes results in another request,
and then fan back in.

This alone ends up demontsrating the kind of complexity cff handles
concurrently.