docs(get-started): Handle error first
@JacobOaks mentioned that before we run `go generate`,
we should have a "good" program that we think will work.
Previously, we were ignoring the error variable
until after fixing the build (with cff.Results).

With this change, we introduce error handling first,
then we fix the build with cff.Results
and add the print statement to consume the response in the process.