make: Fix pwd for Windows
`$(shell pwd)` doesn't work on Windows as-is.
Use MAKEFILE_LIST to get the path to the Makefile,
and then work back to find the directory from that.
This is cross-platform.

Also make the Makefile always run the bash shell,
so we it doesn't try to powershell anything.

Lastly, make the cff, mockgen, staticcheck, and mdox targets
relative to the directory rather than absolute.
Their absolute paths aren't necessary; bin is added to the PATH
so just their names can be used directly.