ci: Run 'make generate' with coverage
This adds a 'make generate-cover' target that
uses Go 1.20's support for coverage-instrumented binaries
to build a 'cff' executable that tracks coverage.
It then runs 'make generate' with that binary,
and collects the coverage data produced by it.

With this change, we'll include coverage of 'make generate'
in the coverage data for the project.

In the Makefile, this makes use of `?=`
to override the GOBIN and BUILD_FLAGS
without creating too much fuss or duplication otherwise.

Resolves #57