ci: Set up dependabot
Sets up dependabot to update:

- GitHub Actions
- Go dependencies in tools/
- Node dependencies in docs/

Go dependencies for the project itself are not upgraded regularly
because it's partly a library, so doing so would force all users
to always upgrade their dependencies.

For Node, because the ecosystem publishes new releases more often,
we're going to upgrade monthly instead of weekly.
(A patch release every week can get annoying.)